requirements.txt
flask==2.0.1
numpy==1.22.4
opencv-python-headless==4.5.5.64
pillow==9.5.0
pytesseract==0.3.10
gunicorn==20.1.0
Werkzeug==2.0.2